SaveTheLibrary: fix Review Book List
